Added function that "slowly" enables physics simulation
I noticed that there was a strange issue if I tried to dynamically enable physics simulation during the game;  For some reason, the tower would instantly fall apart, as if I added an impulse/velocity to the blocks.

To fix this, I created a function that sets the linear damping of all of the blocks to 100.0 (this value seemingly prevents the blocks from moving at all when physics sim is enabled), enables physics simulation on all 54 blocks, and slowly (over the course of 1.5 seconds) ramps down the linear damping of each of the 54 blocks to 0.01 (their default value).

While this is certainly not a perfect solution, this allows me to dynamically enable/disable physics simulation without worrying that the tower will instantly explode/fall apart when the setting is changed.